refactor: modularize analysis logic and rename entry point to index.js
- Split content loading and analysis logic into separate files
- Replace analysisService.js with new index.js as main entry
