--- src/services/analysisService.js
@@ -1,61 +1,0 @@
-import fs from 'fs';
-import { createRequire } from 'module';
-import puppeteer from 'puppeteer';
-
-const require = createRequire(import.meta.url);
-const axeScriptContent = fs.readFileSync(require.resolve('axe-core'), 'utf-8');
-
-export const analyzeAccessibility = async ({ url, htmlContent }) => {
-  let browser;
-  try {
-    browser = await puppeteer.launch({
-      headless: true,
-      args:
-        process.env.NODE_ENV === 'production' ? ['--no-sandbox', '--disable-setuid-sandbox'] : [],
-    });
-    const page = await browser.newPage();
-
-    if (url) {
-      const urlObj = new URL(url);
-
-      if (!['http:', 'https:'].includes(urlObj.protocol)) {
-        throw new Error('HTTP와 HTTPS만 허용됩니다.');
-      }
-
-      await page.goto(url, { waitUntil: 'networkidle0' });
-    } else if (htmlContent) {
-      if (typeof htmlContent !== 'string' || htmlContent.length > 1000000) {
-        throw new Error('유효하지 않은 HTML 콘텐츠가 제공됐습니다.');
-      }
-
-      await page.setContent(htmlContent, { waitUntil: 'networkidle0' });
-    } else {
-      throw new Error('URL 또는 HTML 콘텐츠 중 하나는 반드시 제공되어야 합니다.');
-    }
-
-    try {
-      await page.addScriptTag({ content: axeScriptContent });
-    } catch (error) {
-      throw new Error(`axe-core script 주입 실패: ${error.message}`);
-    }
-
-    const results = await page.evaluate(() => {
-      if (typeof axe === 'undefined') {
-        throw new Error('axe-core를 사용할 수 없습니다.');
-      }
-
-      const options = { runOnly: { type: 'tag', values: ['wcag2a', 'wcag2aa', 'best-practice'] } };
-
-      return axe.run(document, options);
-    });
-
-    return results;
-  } catch (error) {
-    console.error('접근성 분석 중 오류 발생:', error);
-    throw new Error(`분석 실패: ${error.message}`);
-  } finally {
-    if (browser) {
-      await browser.close();
-    }
-  }
-};
